chore(commands): moving commands out of extension.activate
